+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/AutoResizableTextareaMobileOnTouchStart.test.jsx
@@ -3,7 +3,15 @@ import React from 'react';
 import Textarea from '../../../Textarea';
 
 export default class textarea extends React.Component {
-  componentWillMount() {
+  constructor() {
+    super();
+    if (!window.ontouchstart) {
+      this.resetontouchstart = true;
+      window.ontouchstart = 'true';
+    }
+  }
+
+  componentDidUpdate() {
     if (!window.ontouchstart) {
       this.resetontouchstart = true;
       window.ontouchstart = 'true';